Upgrade program, change types

+++ Lab_01/object.hpp
@@ -3,7 +3,7 @@
 
 class TObject {
     public:
-        unsigned long long key;
+        uint64_t key;
         std::string value;
 
         // constructors
